fix(audit): 修复审计报告中的四项问题
- codegen：实体与 DAO 输出改用 UTF-8 写入，修复 Windows 非 UTF-8 环境下中文注释与列名乱码
- codegen：Types.FLOAT 映射由 Float 改为 Double，避免高精度列被截断；仅 Types.REAL 保留 Float
- starter：默认 JdbcTemplateHelper Bean 增加 @ConditionalOnMissingBean，允许使用者自定义覆盖
- page：新增顶层 SELECT 定位工具，修复 GBase8s / Informix 分页在前导空白、注释、CTE 场景生成非法 SQL
- 补充 codegen 编码与类型映射、GBase8s / Informix 分页回归测试
- AUDIT_REPORT.md 同步标注上述四项为已修复并记录验证结果

验证：tiny-jdbc-core 全量 176 个测试通过；tiny-jdbc-codegen 13 个测试通过；tiny-jdbc-boot-starter 编译通过

+++ tiny-jdbc-codegen/src/test/java/org/tinycloud/jdbc/codegen/util/TypeUtilsTest.java
@@ -36,7 +36,7 @@ public class TypeUtilsTest {
     public void testBooleanAndFloatMappings() {
         assertEquals("Boolean", TypeUtils.getJavaType(Types.BOOLEAN, 1, 0));
         assertEquals("Boolean", TypeUtils.getJavaType(Types.BIT, 1, 0));
-        assertEquals("Float", TypeUtils.getJavaType(Types.FLOAT, 12, 2));
+        assertEquals("Double", TypeUtils.getJavaType(Types.FLOAT, 12, 2));
         assertEquals("Float", TypeUtils.getJavaType(Types.REAL, 12, 2));
         assertEquals("Double", TypeUtils.getJavaType(Types.DOUBLE, 22, 2));
     }
